Move classes de database para clarice
